Simple Calculator: Console Input
The aim of this stage was to take the user input and store it in R4. The value was entered by the user as a string of ASCII characters e.g if the user enters “22” it will be stored as 16 (hexadecimal) in R4. My program reads the number entered by the user and then compares it to the return key (0x0D). If the number is equal the while loop ends. If the number is not the same as the return key the program multiplies the number entered by the user by 10 (if it is greater than a single digit number) and converts the ASCII character to numeric by subtracting the ASCII value.