Expression Evaluation
The aim of part two of this assignment was to read and evaluate the expressions entered by the user and store the result in R5, e.g if the user enters “22+17”, using my program from part one, “22” will be stored as 16 (in hexadecimal) in R4 while “17” will be stored as “11” in R9. The sum of these two numbers (39) will be stored in R5 as 27. 
Continuing on from part one, the program compares R0 to the values of 0 and 9 (0x30 and 0x39). This checks if the user input is a number. If the user input is a number the program then goes on to read the next number and stores the value in R9. The program then reads the operator, comparing R7 to “+” (0x2B), “-“ (0x2D), “*” (0x2A) and “/” (0x2F). Finally, if R7 is equal to any of these the program branches to the appropriate operator, computes these expressions and stores the values in R5. I attempted to divide but was unsuccessful, possibly a mistake in the with the adding the count.