Displaying the Result
The aim of part three of this assignment was to display the result stored in R5 in the console window using the program from part 2. I was unable to get my program to do this correctly, instead of displaying the correct result it just displays 0. This is possibly due to the use of branches or an incorrect register.
The program compares the value stored in R11 (9) to the value stored in R10 (0). It then calculates 10^9 and counts how many times 10 must be subtracted on order to prevent the number from getting too large, this is done by adding 1 to the R10 each time. This is repeated until the correct answer is reached.