Symmetric Difference
This program takes two mathematical sets, A and B, containing 32-bit unsigned integers stored in memory. It then attempts to make a third set, C, that consists of the symmetric difference between A and B. I start by loading the values from ASize, AElems, BSize, Csize and CSize. I then take one element in A and compare it to each element in B. If a match is found it branches to match where is stores a ‘$’ in BElems to delete the matching character, it also moves onto the next address in AElems, adds one to the count and reloads BElems so all values are tested against each value in AElems. If no match is found the program stores the values from AElems in CElems, adds one to the count for C and stores this count in CSize. It then moves to the next address in BElems and adds one to the count. Once ASize is equal to the count the program branches to the end.