Countdown Checker
This program looks at two strings and determines if one of those strings can be formed from the other, e.g “beets” can be formed from "daetebzsb". If the word can be formed it stores 1 in R0 and 0 if it can’t be formed. This program compares one letter in cdWord to all the letters in cdLetters, once a match is found it moves on to the next letter in cdWord. I start by loading in the values of the two strings, a count for cdWord and a count for ‘$’. I first work out how many letters are in cdWord using the count, this will then be compared to the count for ‘$’ at the end. Both strings are null terminated, so the program compares cdWord with 0 and if they equal it branches to endWhl where it compares the two counts, if they equal the word can be formed and if not then the word can’t be formed. It then compares the letters between cdWord and cdLetters, if they match it branches to match where it moves to the next address in cdWord and replaces the matching letter with a ‘$’. It then reloads cdLetters so the next letter in cdWord is compared to all the letters in cdLetters. Alternatively, if the letters did not match the program would just move onto the next address in cdLetters and then compares it to 0 as this is also null terminated. Depending on the outcome (i.e. either it is null terminated, and the word can’t be formed, or it can be) or it branches to notFormed or formed where the program either stores 1 in R0 (if the word can be formed) or 0 in R0 (if the word can’t be formed).