Lottery
In the weekly lottery draw for the imaginary lottery company, players can choose six numbers between 1 and 32. They win if they match 4, 5 or 6 numbers. This program determines the number of tickets that have 4 matches, 5 matches, or 6 matches. The results are stored in MATCH4, MATCH5, and MATCH6. This program starts by loading the values for TICKETS, DRAW and COUNT as well as setting various counts. It compares the number position in the ticket with 0 and branches to loop if equal, where it compares the number of matches and resets the counts while moving onto the next address in TICKETS. It then compares the COUNT with the number of times the program has gone through the tickets. If these equal if branches to displayAns which stores the results in the corresponding registers. Finally, it compares the amount of numbers matched with the ticket size, if these equal it branches to while where it reloads DRAW and moves onto the next address in TICKETS. The program compares one number in TICKETS with all the numbers in DRAW and if it finds a match or has exceeded the ticket size it moves onto the next number in TICKETS and begins to compare it.